+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Unreleased
   by default)
 - Added support for usage of perf map files as part of process symbolization
   - Added `perf_map` attribute to `symbolize::Process` type
+- Made `inspect::SymType` enum non-exhaustive
 
 
 0.2.0-alpha.10